refact: split parse_frame and extract the bufferable part out
